+++ example.config.yml
@@ -3,6 +3,8 @@ logging: False
 #日志"
 port: 17860
 #webui 默认启动端口号"
+key:
+#webui 默认token认证key"
 library: 
      strategy: "sogowx:3 bingsite:2 rtst:2 agents:0"
      #知识库参数，每组参数间用空格分隔，冒号前为知识库类型，后为抽取数量。
