+++ views/templates/index.html
@@ -13,14 +13,14 @@
 	<script src="https://npm.elemecdn.com/vue@2.x/dist/vue.min.js"></script>
 	<script src="https://npm.elemecdn.com/vuetify@2.x/dist/vuetify.min.js"></script> -->
 
-	<link href="static/mdi_font/css/materialdesignicons.min.css" rel="stylesheet">
+	<link href="/static/mdi_font/css/materialdesignicons.min.css" rel="stylesheet">
 
-	<link href="static/vuetify.min.css" rel="stylesheet">
-	<script src="static/vue.js"></script>
-	<script src="static/vue-i18n.js"></script>
-	<script src="static/vuetify.min.js"></script>
-	<script src="static/markdown-it.min.js"></script>
-	<script src="static/strings.js"></script>
+	<link href="/static/vuetify.min.css" rel="stylesheet">
+	<script src="/static/vue.js"></script>
+	<script src="/static/vue-i18n.js"></script>
+	<script src="/static/vuetify.min.js"></script>
+	<script src="/static/markdown-it.min.js"></script>
+	<script src="/static/strings.js"></script>
 	<style>
 		.v-navigation-drawer__content::-webkit-scrollbar {
 			width: 5px;
@@ -745,8 +745,8 @@
 			app.chat.push({ "role": role, "content": content })
 		}
 	</script>
-	<script src="api/llm"></script>
-	<script src="api/plugins"></script>
+	<script src="/api/llm"></script>
+	<script src="/api/plugins"></script>
 </body>
 
 </html>
